requirements.txt
Django>=5.2
djangorestframework>=3.16
django-cors-headers>=4.9.0
drf-yasg>=1.21.0
mysqlclient
bcrypt